bots/WORKER-GoogleMapsScraper/requirements.txt
playwright>=1.38
geopy>=2.3
pandas>=2.0
openpyxl>=3.1
psycopg2-binary>=2.9
